requirements.txt
asttokens==2.2.1
attrs==23.1.0
backcall==0.2.0
certifi==2023.7.22
cffi==1.15.1
chromedriver==2.24.1
colorama==0.4.6
comm==0.1.4
cryptography==41.0.3
debugpy==1.6.7.post1
decorator==5.1.1
et-xmlfile==1.1.0
exceptiongroup==1.1.2
executing==1.2.0
h11==0.14.0
idna==3.4
importlib-metadata==6.8.0
install==1.3.5
ipykernel==6.25.1
ipython==8.14.0
jedi==0.19.0
jupyter_client==8.3.0
jupyter_core==5.3.1
matplotlib-inline==0.1.6
nest-asyncio==1.5.7
numpy==1.25.2
openpyxl==3.1.2
oracledb==1.3.2
outcome==1.2.0
packaging==23.1
pandas==2.0.3
parso==0.8.3
pickleshare==0.7.5
platformdirs==3.10.0
prompt-toolkit==3.0.39
psutil==5.9.5
pure-eval==0.2.2
pycparser==2.21
PyGetWindow==0.0.9
Pygments==2.16.1
pyodbc==4.0.39
PyRect==0.2.0
PySocks==1.7.1
python-dateutil==2.8.2
python-dotenv==1.0.0
pytz==2023.3
pywin32==306
pyzmq==25.1.0
selenium==4.11.2
six==1.16.0
sniffio==1.3.0
sortedcontainers==2.4.0
stack-data==0.6.2
tornado==6.3.2
traitlets==5.9.0
trio==0.22.2
trio-websocket==0.10.3
typing_extensions==4.7.1
tzdata==2023.3
urllib3==2.0.4
watchdog==3.0.0
wcwidth==0.2.6
wsproto==1.2.0
XlsxWriter==3.1.2
zipp==3.16.2
